This new branch allows the window to be resized and changes the maximum/minimum x/y values accordingly.
LowpanNode needs to be sent the ui that it will be drawn on in the constructor, otherwise it will not know what it can set its location to

NodeCanvas calls forceLocationOnCanvas before drawing the nodes. This is optional and does mess with the shape of the network when nodes would go off screen.